tests/wpt/tests/css/css-animations/parsing/animation-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: animation sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation">
<meta name="assert" content="animation supports the full grammar '<single-animation> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('animation', 'anim paused both reverse 4 1s -3s cubic-bezier(0, -2, 1, 3)', {
  'animation-duration': '1s',
  'animation-timing-function': 'cubic-bezier(0, -2, 1, 3)',
  'animation-delay': '-3s',
  'animation-iteration-count': '4',
  'animation-direction': 'reverse',
  'animation-fill-mode': 'both',
  'animation-play-state': 'paused',
  'animation-name': 'anim',
  'animation-timeline': 'auto',
  'animation-range-start': 'normal',
  'animation-range-end': 'normal',
});

test_shorthand_value('animation', 'anim paused both reverse, 4 1s -3s cubic-bezier(0, -2, 1, 3)', {
  'animation-duration': 'auto, 1s',
  'animation-timing-function': 'ease, cubic-bezier(0, -2, 1, 3)',
  'animation-delay': '0s, -3s',
  'animation-iteration-count': '1, 4',
  'animation-direction': 'reverse, normal',
  'animation-fill-mode': 'both, none',
  'animation-play-state': 'paused, running',
  'animation-name': 'anim, none',
  'animation-timeline': 'auto',
  'animation-range-start': 'normal',
  'animation-range-end': 'normal',
});

test_shorthand_value('animation', '4 1s -3s cubic-bezier(0, -2, 1, 3), anim paused both reverse', {
  'animation-duration': '1s, auto',
  'animation-timing-function': 'cubic-bezier(0, -2, 1, 3), ease',
  'animation-delay': '-3s, 0s',
  'animation-iteration-count': '4, 1',
  'animation-direction': 'normal, reverse',
  'animation-fill-mode': 'none, both',
  'animation-play-state': 'running, paused',
  'animation-name': 'none, anim',
  'animation-timeline': 'auto',
  'animation-range-start': 'normal',
  'animation-range-end': 'normal',
});
</script>
</body>
</html>
